适配代码
