--- Esp32/config.json
@@ -1,13 +1,0 @@
-{
-    "wifi": {
-        "ssid": "",
-        "password": ""
-    },
-    "baidu": {
-        "api_key": "",
-        "secret_key": ""
-    },
-    "silicon": {
-        "api_token": ""
-    }
-}
