+++ Pi/MAIN/mood.py
@@ -1,12 +1,8 @@
-# 迁移自Esp32/mood.py
-# 需检查MicroPython相关内容并适配为标准Python
-# 其余功能结构和流程保持不变
-
 import requests
 import json
 import asyncio
 
-SERVER_URL = 'http://192.168.101.246:5001/api/mood'
+SERVER_URL = 'http://127.0.0.1:5001/api/mood'
 
 async def send_mood(mood):
     headers = {'Content-Type': 'application/json'}
