--- Pi/MAIN/wifi_connect.py
@@ -1,33 +1,0 @@
-# 迁移自Esp32/wifi_connect.py
-# 需检查MicroPython相关内容并适配为标准Python
-# 其余功能结构和流程保持不变
-
-# ... existing code ... 
-
-# 树莓派系统通常自动联网，无需代码连接WiFi
-# 可选：通过iwconfig或nmcli获取WiFi信号质量
-import subprocess
-
-def get_wifi_quality(interface='wlan0'):
-    try:
-        # 使用iwconfig获取信号强度
-        result = subprocess.run(['iwconfig', interface], capture_output=True, text=True)
-        output = result.stdout
-        if 'Link Quality' in output:
-            # 解析Link Quality=xx/yy
-            import re
-            match = re.search(r'Link Quality=(\d+)/(\d+)', output)
-            if match:
-                quality = int(match.group(1)) / int(match.group(2)) * 100
-                return int(quality)
-        return None
-    except Exception as e:
-        print(f'获取WiFi质量失败: {e}')
-        return None
-
-async def do_connect(WIFI_SSID, WIFI_PASSWORD):
-    print('树莓派系统已自动联网，无需手动连接WiFi')
-    return True
-
-if __name__ == "__main__":
-    print(get_wifi_quality())
